+++ changelog.md
@@ -5,6 +5,7 @@
 * Added missing `bind()` calls on `Ads` plugin @ron666
 * Fixed issue with `VAST` plugin due to updates on AJAX method @ron666
 * Integrated VPAID 2.0 with `VAST` plugin @ron666  
+* Integrated events to load tracking URLs in `VAST` plugin @ron666
 
 *2.0.0 (2017/03/22)*